refactor

+++ modules/pi_estimator.py
@@ -9,4 +9,4 @@ def python_estimate_pi(samples: int) -> float:
     x = np.random.uniform(-1.0, 1.0, samples)
     y = np.random.uniform(-1.0, 1.0, samples)
     inside_circle = (x**2 + y**2) <= 1.0
-    return 4.0 * np.sum(inside_circle) / samples
+    return float(4.0 * np.sum(inside_circle) / samples)
